add timer counter in get buslines by busStop id response

+++ domain/models/busline.go
@@ -3,8 +3,8 @@ package models
 type (
 	BusStop struct {
 		Id   string  `json:"id" bson:"id"`
-		Lat  float32 `json:"lat" bson:"lat"`
-		Lng  float32 `json:"lng" bson:"lng"`
+		Lat  float64 `json:"lat" bson:"lat"`
+		Lng  float64 `json:"lng" bson:"lng"`
 		Name string  `json:"name" bson:"name"`
 	}
 
@@ -13,7 +13,7 @@ type (
 		FullName  string      `json:"fullName" bson:"fullName"`
 		Id        string      `json:"id" bson:"id"`
 		Origin    string      `json:"origin" bson:"origin"`
-		Path      [][]float32 `json:"path" bson:"path"`
+		Path      [][]float64 `json:"path" bson:"path"`
 		ShortName string      `json:"shortName" bson:"shortName"`
 	}
 )
